appended notifier
