+++ configs/server/forseti_conf_server.yaml.in
@@ -223,6 +223,8 @@ notifier:
         - resource: iam_policy_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -241,6 +243,8 @@ notifier:
         - resource: audit_logging_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -251,6 +255,8 @@ notifier:
         - resource: blacklist_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -261,6 +267,8 @@ notifier:
         - resource: bigquery_acl_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -271,6 +279,8 @@ notifier:
         - resource: buckets_acl_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -281,6 +291,8 @@ notifier:
         - resource: cloudsql_acl_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -291,6 +303,8 @@ notifier:
         - resource: enabled_apis_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -301,6 +315,8 @@ notifier:
         - resource: firewall_rule_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -311,6 +327,8 @@ notifier:
         - resource: forwarding_rule_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -321,6 +339,8 @@ notifier:
         - resource: ke_version_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -331,6 +351,8 @@ notifier:
         - resource: ke_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -341,6 +363,8 @@ notifier:
         - resource: groups_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -351,6 +375,8 @@ notifier:
         - resource: instance_network_interface_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -361,6 +387,8 @@ notifier:
         - resource: iap_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -371,6 +399,8 @@ notifier:
         - resource: lien_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -381,6 +411,8 @@ notifier:
         - resource: location_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -391,6 +423,8 @@ notifier:
         - resource: log_sink_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -401,6 +435,8 @@ notifier:
         - resource: service_account_key_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -411,6 +447,8 @@ notifier:
         - resource: external_project_access_violations
           should_notify: true
           notifiers:
+            # Email violations
+            - name: email_violations
             # Upload violations to GCS.
             - name: gcs_violations
               configuration:
@@ -444,6 +482,4 @@ notifier:
         gcs_path: gs://{FORSETI_BUCKET}/inventory_summary
       email_summary:
         enabled: true
-        sendgrid_api_key: {SENDGRID_API_KEY}
-        sender: {EMAIL_SENDER}
-        recipient: {EMAIL_RECIPIENT}
+
